back-end/pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.0",
    "fastapi>=0.135.3",
    "inngest>=0.5.18",
    "llama-index-core>=0.14.20",
    "llama-index-readers-file>=0.6.0",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
    "python-multipart>=0.0.26",
]
